+++ src/leveldb/db/log_reader.cc
@@ -2,12 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 
-#include "db/log_reader.h"
+#include "leveldb/db/log_reader.h"
 
 #include <stdio.h>
 #include "leveldb/env.h"
-#include "util/coding.h"
-#include "util/crc32c.h"
+#include "leveldb/util/coding.h"
+#include "leveldb/util/crc32c.h"
 
 namespace leveldb {
 namespace log {
@@ -180,7 +180,7 @@ void Reader::ReportDrop(size_t bytes, const Status& reason) {
 
 unsigned int Reader::ReadPhysicalRecord(Slice* result) {
   while (true) {
-    if (buffer_.size() < kHeaderSize) {
+    if (buffer_.size() < (int)kHeaderSize) {
       if (!eof_) {
         // Last read was a full read, so this is a trailer to skip
         buffer_.clear();
@@ -191,7 +191,7 @@ unsigned int Reader::ReadPhysicalRecord(Slice* result) {
           ReportDrop(kBlockSize, status);
           eof_ = true;
           return kEof;
-        } else if (buffer_.size() < kBlockSize) {
+        } else if (buffer_.size() < (int)kBlockSize) {
           eof_ = true;
         }
         continue;
