+++ src/leveldb/db/memtable.cc
@@ -2,12 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 
-#include "db/memtable.h"
-#include "db/dbformat.h"
+#include "leveldb/db/memtable.h"
+#include "leveldb/db/dbformat.h"
 #include "leveldb/comparator.h"
 #include "leveldb/env.h"
 #include "leveldb/iterator.h"
-#include "util/coding.h"
+#include "leveldb/util/coding.h"
 
 namespace leveldb {
 
@@ -101,7 +101,7 @@ void MemTable::Add(SequenceNumber s, ValueType type,
   p += 8;
   p = EncodeVarint32(p, val_size);
   memcpy(p, value.data(), val_size);
-  assert((p + val_size) - buf == encoded_len);
+  assert((p + val_size) - buf == (int)encoded_len);
   table_.Insert(buf);
 }
