+++ src/leveldb/util/coding.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 
-#include "util/coding.h"
+#include "leveldb/util/coding.h"
 
 namespace leveldb {
 
@@ -81,7 +81,7 @@ void PutVarint32(std::string* dst, uint32_t v) {
 char* EncodeVarint64(char* dst, uint64_t v) {
   static const int B = 128;
   unsigned char* ptr = reinterpret_cast<unsigned char*>(dst);
-  while (v >= B) {
+  while (v >= (uint64_t)B) {
     *(ptr++) = (v & (B-1)) | B;
     v >>= 7;
   }
